refactor UserService.checkIfUserHasPermissionToAlter - method now checks if user either is admin or is the owner. Delete duplicated checks from AdsServiceImpl & CommentServiceImpl; add new UserHasNoRightsException;

+++ src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
@@ -90,7 +90,7 @@ public class AdsServiceImpl implements AdsService {
         log.info("Was invoked removeAds method from {}", AdsService.class.getSimpleName());
         Ads ads = getAdsById(id);
 
-        checkIfUserCanAlterAds(authentication, ads);
+        userService.checkIfUserHasPermissionToAlter(authentication, ads.getAuthor().getEmail());
         adsRepository.delete(ads);
     }
 
@@ -106,8 +106,7 @@ public class AdsServiceImpl implements AdsService {
     public AdsDto updateAdsById(long id, CreateAdsDto createAdsDto, Authentication authentication) {
         log.info("Was invoked updateAdsById method from {}", AdsService.class.getSimpleName());
         Ads oldAds = getAdsById(id);
-
-        checkIfUserCanAlterAds(authentication, oldAds);
+        userService.checkIfUserHasPermissionToAlter(authentication, oldAds.getAuthor().getEmail());
         Ads infoToUpdate = adsMapper.createAdsDtoToAds(createAdsDto);
 
         oldAds.setPrice(infoToUpdate.getPrice());
@@ -116,16 +115,6 @@ public class AdsServiceImpl implements AdsService {
 
         Ads updatedAds = adsRepository.save(oldAds);
         return adsMapper.adsToAdsDto(updatedAds);
-    }
-
-    private void checkIfUserCanAlterAds(Authentication authentication, Ads ads) {
-        boolean matchUser = authentication.getName().equals(ads.getAuthor().getEmail());
-        boolean userIsAdmin = userService.checkIfUserIsAdmin(authentication);
-
-        if (!(userIsAdmin || matchUser)){
-            log.warn("Current Ads isn't of authentication user!");
-            throw new RuntimeException("403 Forbidden");
-        }
     }
 
     /**
